+++ __tests__/modules/req.test.ts
@@ -266,6 +266,7 @@ describe('Request extensions', () => {
           'Content-Type': 'text/html; charset=UTF-8',
         },
       })
+
       expect.assertions(1)
     })
   })