+++ examples/typescript-decorators/index.ts
@@ -4,7 +4,7 @@ import { basePath, get, register } from 'express-decorators'
 @basePath('/hello')
 class TestController {
   @get('/world')
-  async sayHelloAction(_: Request, res: Response) {
+  sayHelloAction(_: Request, res: Response) {
     res.send('<h1>Hello World</h1>')
   }
 }