+++ packages/res/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @tinyhttp/res
 
+## 0.4.9
+
+### Patch Changes
+
+- Fix markdown recursive issue
+- Updated dependencies [undefined]
+  - @tinyhttp/req@0.4.5
+
+## 0.4.8
+
+### Patch Changes
+
+- @tinyhttp/req@0.4.4
+
 ## 0.4.7
 
 ### Patch Changes